==> scripts/05_final_report.Rmd <==
---
title: "Perceived Risk and Marijuana Use Among Adolescents (NSDUH 2023)"
student: Ganyu Liu
output:
  word_document:
    toc: true
  pdf_document:
    toc: true
  html_document:
    theme: readable
    toc: true
    toc_float: true
    number_sections: true
editor_options:
  markdown:
    wrap: 72
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(
  echo = TRUE,
  warning = FALSE,
  message = FALSE
)

library(tidyverse)
library(car)
library(odds.n.ends)
library(ggplot2)
library(DiagrammeR)
library(DiagrammeRsvg)
library(rsvg)
library(table1)
```

# Background

Adolescent marijuana use is a persistent public health issue. Although use has remained relatively stable, perceived harm from marijuana continues to decrease among youth. Understanding how perceived risk relates to actual use can help inform prevention strategies and targeted public health messaging.
This analysis uses the 2023 National Survey on Drug Use and Health (NSDUH) to examine the relationship between perceived risk of weekly marijuana use and past 30-day marijuana use among adolescents ages 12–17. We also explore whether this relationship differs by sex.

# Methods

## Data Source
Data come from the 2023 NSDUH public-use file. All data cleaning, recoding, and analytic preparation were completed systematically and using reproducible scripts. 

## Key Variables 
Exposure: Perceived risk of smoking marijuana weekly (RSKMRJWK) 
- Categories: No risk, Slight risk, Moderate risk, Great risk 
Outcome: Past 30-day marijuana use (MJDAY30A)
- Categories: No use, Used 
Covariate: Sex (IRSEX: Male, Female) 

## Statistical Analysis 
Descriptive statistics summarized the sample and the distribution of exposure, outcome, and covariates. Logistic regression was used to estimate the association between perceived risk and marijuana use. Three models were fit: Unadjusted model Adjusted
model including sex Interaction model testing risk × sex Predicted probabilities and ROC curves to assess model performance.

## Study Population 
The sample included adolescents ages 12–17 with complete information on perceived risk, past 30-day marijuana use, and sex. Exclusion steps were applied. A flow diagram (Figure 1) summarizes how the final analytic dataset was created. 

# Results 
## Study Flow and Exclusion Diagram
```{r}
# Load raw NSDUH data
load("~/Desktop/FA2025/6009/assignment-1-RQ/Github/FP3_NSDUH_AdolescentCannabis/data_raw/NSDUH_2023.Rdata")

# The raw object is named "data"
data_raw <- data

# Step 0: Select only variables used in cleaning (same as 01 script)
data_selected <- data_raw %>%
  select(RSKMRJWK, MJDAY30A, IRSEX, AGE3,
         YEPMJMO, YFLMJMO, NEWRACE2)

# Step 1: Adolescents only (correct code)
data_adol <- data_selected %>%
  filter(AGE3 %in% c(1,2,3))

start_n <- nrow(data_adol)

# Step 2: Exclude missing perceived risk (same as 01)
ex1 <- data_adol %>%
  mutate(
    RSKMRJWK = as.numeric(RSKMRJWK),
    RSKMRJWK = ifelse(RSKMRJWK %in% c(85,94,97,98), NA, RSKMRJWK)
    ) %>%
  drop_na(RSKMRJWK)

n_ex1 <- start_n - nrow(ex1)

# Step 3: Exclude missing marijuana use (same as 01 recode rules)
ex2 <- ex1 %>%
  mutate(
    MJDAY30A = case_when(
      MJDAY30A %in% c(91,93) ~ 0,
      MJDAY30A %in% 1:30 ~ 1,
      MJDAY30A %in% c(94,97,98) ~ NA_real_
    )
  ) %>%
  drop_na(MJDAY30A)

n_ex2 <- nrow(ex1) - nrow(ex2)

# Step 4: Exclude missing sex
ex3 <- ex2 %>% drop_na(IRSEX)
n_ex3 <- nrow(ex2) - nrow(ex3)

# Final analytic sample size
analytic <- ex3
final_n <- nrow(ex3)

start_n; n_ex1; n_ex2; n_ex3; final_n

# Recode IRSEX properly (1 = Male, 2 = Female)
analytic$IRSEX <- factor(analytic$IRSEX,
                         levels = c(1,2),
                         labels = c("Male","Female"))

# Recode AGE3 properly (1,2,3)
analytic$AGE3 <- as.numeric(analytic$AGE3)

# Clean peer disapproval (YFLMJMO)
analytic <- analytic %>%
  mutate(
    YFLMJMO = as.numeric(YFLMJMO),
    YFLMJMO = case_when(
      YFLMJMO == 1 ~ 1,     # disapprove
      YFLMJMO == 2 ~ 0,     # neutral/permissive
      TRUE ~ NA_real_       # missing
    ),
    YFLMJMO = factor(
      YFLMJMO,
      levels = c(0,1),
      labels = c("Neutral/Permissive", "Disapprove")
    )
  )
# Clean parent disapproval (YEPMJMO)
analytic <- analytic %>%
  mutate(
    YEPMJMO = as.numeric(YEPMJMO),
    YEPMJMO = case_when(
      YEPMJMO %in% c(2,3) ~ 1,                   # disapprove
      YEPMJMO == 1 ~ 0,                          # neutral
      YEPMJMO %in% c(85,94,97,98,99) ~ NA_real_, # missing
      TRUE ~ NA_real_
    ),
    YEPMJMO = factor(
      YEPMJMO,
      levels = c(0,1),
      labels = c("Neutral", "Disapprove")
    )
  )

# Fix perceived risk coding
analytic$RSKMRJWK <- factor(
  as.numeric(analytic$RSKMRJWK),
  levels = c(1,2,3,4),
  labels = c("No risk","Slight risk","Moderate risk","Great risk")
)
```
Figure 1. Flowchart of exclusions

```{r}
p <- grViz("digraph flowchart {

      node [fontname = Helvetica, shape = rectangle, fontsize = 15] 
      
      node1 [label = '@@1']
      node2 [label = '@@2']
      node3 [label = '@@3']
      node4 [label = '@@4']
      
      node1 -> node2 -> node3 -> node4
}
      [1]: 'Adolescents ages 12-17 in NSDUH 2023 (N = 11572)'
      [2]: 'Excluding missing perceived risk (RSKMRJWK): n excluded = 305; n remaining = 11267'
      [3]: 'Excluding missing past 30-day marijuana use (MJDAY30A): n excluded = 115; n remaining = 11152'
      [4]: 'Excluding missing sex (IRSEX): n excluded = 0; Final analytic sample = 11152'
      ")
p
```
## DAG 

```{r}
library(dagitty)

dag <- dagitty("
dag {
  PerceivedRisk
  MarijuanaUse
  Age
  Sex
  PeerDisapproval
  ParentDisapproval

  PerceivedRisk -> MarijuanaUse

  Age -> PerceivedRisk
  Age -> MarijuanaUse

  Sex -> PerceivedRisk
  Sex -> MarijuanaUse

  PeerDisapproval -> PerceivedRisk
  PeerDisapproval -> MarijuanaUse

  ParentDisapproval -> PerceivedRisk
  ParentDisapproval -> MarijuanaUse
}
")

coordinates(dag) <- list(
  x = c(
    Age = 0.1,
    Sex = 0.1,
    PeerDisapproval = 0.1,
    ParentDisapproval = 0.1,
    PerceivedRisk = 0.5,
    MarijuanaUse = 0.9
  ),
  y = c(
    Age = 0.8,
    Sex = 0.2,
    PeerDisapproval = 0.6,
    ParentDisapproval = 0.4,
    PerceivedRisk = 0.5,
    MarijuanaUse = 0.5
  )
)
plot(dag, lwd = 2)
plot(dag)
```

## Descriptive Cross-Tabulations

```{r}
table(analytic$RSKMRJWK)
table(analytic$MJDAY30A)
table(analytic$IRSEX)

risk_use <- table(analytic$RSKMRJWK, analytic$MJDAY30A)
risk_use

round(prop.table(risk_use, 1) * 100, 1)
```
Interpretation: 
- Use prevalence decrease from 24% among adolescents reporting no perceived risk to under 1% among those reporting great risk.
- Crude cross-tab shows a strong inverse does-responde pattern.
- These patterns support the hypothesis that lower perceived risk is associated with higher past 30 days use. 

### Table 1. Descriptive Characteristics
```{r}
library(table1)

label(analytic$AGE3)     <- "Age (3-level)"
label(analytic$IRSEX)    <- "Sex"
label(analytic$RSKMRJWK) <- "Perceived risk of weekly marijuana use"
label(analytic$YFLMJMO)  <- "Peer disapproval of marijuana use"
label(analytic$YEPMJMO)  <- "Parental disapproval of marijuana use"
label(analytic$MJDAY30A) <- "Past 30-day marijuana use"

# Table 1.
table1(
  ~ AGE3 + IRSEX + RSKMRJWK + YFLMJMO + YEPMJMO + MJDAY30A,
  data         = analytic,
  overall      = "Total",
  rowlabelhead = "Variable",
  caption      = "Table 1. Descriptive Characteristics of the Analytic Sample"
)
```

## Figure 2. Marijuana Use by Perceived Risk

```{r}
# Ensure MJDAY30A is a factor
analytic$MJDAY30A <- factor(analytic$MJDAY30A,
                            levels = c(0,1),
                            labels = c("No use","Any use"))

ggplot(analytic, aes(x = RSKMRJWK, fill = MJDAY30A)) +
  geom_bar(position = "fill") +
  labs(
    title = "Marijuana Use by Perceived Risk (Adolescents, NSDUH 2023)",
    x = "Perceived Risk",
    y = "Proportion",
    fill = "Past 30-Day Use"
  ) +
  theme_bw()
```
Interpretation: 
This figure 2. shows a strong inverse relationship between perceived risk and past 30-day marijuana use among adolescents. 
In the NO RISK group, 322 adolescents reported marijuana use, compared to 1006 who did not use. 
In the SLIGHT RISK group, 331 reported use and 2447 did not use.
In the MODERATE RISK group, 116 reported use and 3442 did not use.
In the GREAT RISK group, 30 adolescents reported use and 3478 did not use.
As the figure 1. showing, the number of users steadily decreased as perceived risk increased which supports the expected crude association, adolescents who view marijuana as less risk are more likely to have used in the past 30 days.

## Logistic Regression Models

Model 1: Unadjusted

```{r}
model1 <- glm(MJDAY30A ~ RSKMRJWK,
              data = analytic,
              family = "binomial")
summary(model1)
odds.n.ends(model1)
```
Intepretation:
Adolescents who perceived slight, moderate, or great risk of marijuana use had lower odds of reporting past 30-day marijuana use compared with those reporting no risk (reference).
- the odds of past 30 days marijuana use among adolescents who perceive slight risk are 0.42 times the odds among those who report no risk.
- the odds of past 30 days marijuana use among adolescents who perceive moderate risk are 0.11 times the odds among those who report no risk.
- the odds of past 30 days marijuana use among adolescents who perceive great risk are 0.027 times the odds among those who report no risk.
The overall model has p value less than 0.05 which means perceived risk strongly predicts marijuana use.
The correct classification rate is 92.8%, due to the low prevalence of use.

Model 2: Adjusted for Age and Sex

```{r}
model2 <- glm(MJDAY30A ~ RSKMRJWK + AGE3 + IRSEX,
              data = analytic,
              family = "binomial")
summary(model2)
odds.n.ends(model2)
```

Intepretation:
After adjusting for age and sex, perceived risk remains a strong independent predictor, and age shows a major driver of use. Female also show higher reported use than males. 
- The odds of use for adolescents perceiving slight risk are 0.40 times the odds among those perceiving no risk, after adjusting for age and sex.
- The odds of use for adolescents perceiving moderate risk are 0.11 times the odds among those perceiving no risk, after adjusting for age and sex.
- The odds of use for adolescents perceiving great risk are 0.03 times the odds among those perceiving no risk, after adjusting for age and sex.
- Each one-unit increase in age is associated with a 2.46-fold higher odds of marijuana use.
- Females have 1.58 times the odds of marijuana use compared with males.

Likelihood Ratio Test

```{r}
lmtest::lrtest(model1, model2)
```
As the results shown, adding age and sex significantly improves model fit (p-value < 0.05). These covariates explain variation in marijuana use beyond perceived risk alone.

Model 3: Full Adjusted (demographics + peer + parent)

```{r}
complete_data <- analytic %>%
  mutate(
    MJDAY30A = factor(MJDAY30A, levels = c("No use", "Any use"))
  ) %>%
  dplyr::select(MJDAY30A, RSKMRJWK, AGE3, IRSEX, YFLMJMO, YEPMJMO) %>%
  drop_na()
# check size
nrow(complete_data)

model3 <- glm(MJDAY30A ~ RSKMRJWK + AGE3 + IRSEX + YFLMJMO + YEPMJMO,
              data = complete_data,
              family = "binomial")
summary(model3)
odds.n.ends(model3)
```
Intepretation: 
Peer disapproval has a stronger protective association than parental disapproval.
- The odds of use for adolescents perceiving slight risk are 0.55 times the odds among those perceiving no risk, after adjusting for age, sex, peer disapproval, and parental disapproval.
- The odds of use for adolescents perceiving moderate risk are 0.25 times the odds among those perceiving no risk, after adjusting for age, sex, peer disapproval, and parental disapproval.
- The odds of use for adolescents perceiving great risk are 0.08 times the odds among those perceiving no risk, after adjusting for age, sex, peer disapproval, and parental disapproval.
- Each one-unit increase in age is associated with a 2-fold higher odds of marijuana use.
- Females have 1.42 times the odds of marijuana use compared with males.
- Adolescents who report that their friends disapprove of marijuana use have 0.19 times the odds of using, compared with those whose peers are neutral/permissive.
- Adolescents whose parents disapprove have 0.48 times the odds of using compared with those whose parents are neutral.

Likelihood Ratio Test (Adjusted vs Interaction)

```{r}
# before doing the llr test, adjust model 2
model2_adj <- glm(MJDAY30A ~ RSKMRJWK + AGE3 + IRSEX,
                  data = complete_data,
                  family = "binomial")
summary(model2_adj)
odds.n.ends(model2_adj)

lmtest::lrtest(model2_adj, model3)
```
As a result, it clearly showed that adding peer and parental disapproval substantially improves model fit. These are highly important predictors of adolescent marijuana use. 

### Table 2. Adjusted Odds Ratios for Past 30-Day Marijuana Use (Model 3)
```{r}
# Clean variable label map
label_map <- c(
  "RSKMRJWKSlight risk" = "Slight risk vs No risk",
  "RSKMRJWKModerate risk" = "Moderate risk vs No risk",
  "RSKMRJWKGreat risk" = "Great risk vs No risk",
  "AGE3" = "Age (per category)",
  "IRSEXFemale" = "Female vs Male",
  "YFLMJMO" = "Peer disapproval (yes vs no)",
  "YEPMJMO" = "Parental disapproval (yes vs no)",
  "RSKMRJWKSlight risk:IRSEXFemale" = "Interaction: Slight risk × Female",
  "RSKMRJWKModerate risk:IRSEXFemale" = "Interaction: Moderate risk × Female",
  "RSKMRJWKGreat risk:IRSEXFemale" = "Interaction: Great risk × Female"
)

# Table 2
model3_df <- data.frame(
  Variable = rownames(exp(cbind(OR = coef(model3), confint(model3)))),
  exp(cbind(OR = coef(model3), confint(model3)))
)

model3_df <- model3_df[model3_df$Variable != "(Intercept)", ]
model3_df$Variable <- label_map[model3_df$Variable]

colnames(model3_df) <- c("Variable","OR","Lower 95% CI","Upper 95% CI")
model3_df
```

Model 4: Interaction Model (RSKMRJWK X IRSEX)

```{r}
model4 <- glm(MJDAY30A ~ RSKMRJWK * IRSEX + AGE3 + YFLMJMO + YEPMJMO,
              data = complete_data,
              family = "binomial")
summary(model4)
odds.n.ends(model4)
```
Intepretation:
- Interaction terms:
  - For slight risk, females have 0.89 times the odds compared with males at the same perceived risk level. This is not statistically significant due to p-value >0.05.
  - For moderate risk, females have 1.51 times the odds compared with males at the same perceived risk level. This is not statistically significant due to p-value >0.05.
  - For great risk, females have 2.06 times the odds compared with males at the same perceived risk level. This is not statistically significant due to p-value >0.05.
Although perceived risk was strongly associated with lower marijuana use within both males and females, the sex × perceived risk interaction term was not statistically significant.
  
Likehood Ratio Test (full model vs interaction model)

```{r}
lmtest::lrtest(model3, model4)
```
Intepretation:
The interaction model does not significantly outperform Model 3.
Sex does not significantly modify the relationship between perceived risk and marijuana use.

### Table 3. Interaction Model of Perceived Risk × Sex (Model 4)
```{r}
model4_df <- data.frame(
  Variable = rownames(exp(cbind(OR = coef(model4), confint(model4)))),
  exp(cbind(OR = coef(model4), confint(model4)))
)

model4_df <- model4_df[model4_df$Variable != "(Intercept)", ]
model4_df$Variable <- label_map[model4_df$Variable]

colnames(model4_df) <- c("Variable","OR","Lower 95% CI","Upper 95% CI")
model4_df
```

## Assumption checking 
```{r}
library(car)
vif(model3)
```
All values were < 2, indicating no meaningful multicollinearity. 

### MInfluential Observations (Cook’s D, leverage, dfbetas)
```{r}
library(broom)
diag_m3 <- augment(model3) %>%
  mutate(index = 1:n())

# cook's distance plot
plot(model3, which = 4, id.n = 3, col = "pink")
abline(h = 1, col = "pink", lty = 2)

# identify observations above cutoff points (cook's distance)
sum(diag_m3$.cooksd > 1)

# Leverage (hat values)
plot(model3, which = 5, id.n = 3)

# dfbetas summary
summary(dfbetas(model3))

```
Cook’s D, leverage, and dfbetas indicated no influential observations. However, a few cases had slightly higher leverage values, but their Cook’s D values were extremely small (<0.01), showing they might have minimal influence on parameter estimates. No data points required removal.

### Goodness of fit (Hosmer–Lemeshow)

```{r}
library(ResourceSelection)

hoslem.test(
  x = as.numeric(complete_data$MJDAY30A) - 1,  # convert to 0/1
  y = fitted(model3)
)
```
The Hosmer-Lemeshow test was statistically significant (p=0.0025), which is above less than 0.05. This suggests some lack of calibration. The reason could be the sample size is large above 11,000, and the HLT is sensitive to flag any minor deviations, so this would be fine. Visual dianostics (ROC AUC = 0.88) and predicted probability plots both acceptable overall model performance. The model is acceptable fit. 

### Linearity of the logit (AGE3)

```{r}
model3_age_test <- glm(
  MJDAY30A ~ RSKMRJWK + IRSEX + YFLMJMO + YEPMJMO +
    AGE3 + I(AGE3 * log(AGE3)),
  data = complete_data,
  family = binomial
)

summary(model3_age_test)

```
The linearity-in-the-logit assumption applies only to continuous predictors. In this study, age represents three discrete adolescent age groups (12-13, 14-15, 16-17). It functions as an ordinal categorical variable, and therefore the linearity assumption does not apply. A box-tidwell applied, adn reached statistical significance; however, age is not a true continuous measures, this result is not meaningful.
All other predictors were categorical, so the linearity assumption was not relevant, and no violations were detected for variables to which the assumption applies.

## Model Performance

Predicted Probabilities (full adjusted Model)

```{r, fig.width=8, fig.height=6}
odds.n.ends(model3, predProbPlot = TRUE)
```


ROC Curve (full adjusted Model)

```{r}
odds.n.ends(model3, rocPlot = TRUE)
```

Predicted Probabilities (Interaction Model)

```{r, fig.width=8, fig.height=6}
odds.n.ends(model4, predProbPlot = TRUE)
```

ROC Curve (Interaction Model)

```{r}
odds.n.ends(model4, rocPlot = TRUE)
```
Intepretation:
- Most non-using groups had predicted probabilities below 0.10.
- Most using groups had predicted probabilities between 0.10 and 0.60. 
- Overall, this is acceptable because marijuana use is a low proportion in the dataset.
- ROC curves showed strong model discrimination, which models distinguish between users and non-users.
- Low sensitivity is acceptable because the proportion of users in the group is low. Specificity remains high, so ROC curves show strong performance.
- Model 3 and Model 4 show identical predicted probabilities and ROC curves that match the non-significant sex interaction result.

# Discussion

Perceived risk of smoking marijuana once or twice a week was strongly and associated with past 30-day use among adolescents. Youth who viewed marijuana as more harmful had substantially lower odds of recent use, indicating a clear dose–response relationship. This association continued after adjusting for age, sex, peer disapproval, and parental disapproval.
Testing for effect modification showed that the perceived risk–use association was similar for males and females. Although the point estimates differed very slightly, particularly in the moderate- and great-risk categories, the sex × perceived risk interaction was not statistically significant. These findings suggest that perceived risk impacts marijuana use similarly for both male and female adolescents.
Peer and parental disapproval were also strongly associated with lower marijuana use. Adolescents whose friends disapproved of marijuana use had the lowest odds of reporting past-month use, highlighting the major influence of peer norms during adolescence. 

# Reproducibility 
All data cleaning, descriptive analysis, and Modeling scripts are stored in the project repository: 
 - 01_data_cleaning.R 
 - 02_descriptive_analysis.R 
 - 03_modeling.R 
 - 04_logistic_models.R
Workflow follows reproducible research practices with transparent version control and structured outputs.
